udpated trial system integration
