+++ apps/io-wallet-user-func/src/infra/http/handlers/__test__/health.spec.ts
@@ -30,6 +30,7 @@ describe("HealthHandler", () => {
 
   // test di quando questa va in errore
   const trialSystemClient: TrialSystemHealthCheck = {
+    featureFlag: "true",
     healthCheck: () => TE.right(true),
   };
