+++ apps/io-wallet-user-func/src/infra/http/handlers/__test__/set-wallet-instance-status.spec.ts
@@ -36,14 +36,12 @@ describe("SetWalletInstanceStatusHandler", () => {
 
   // test di quando questa va in errore
   const userTrialSubscriptionRepository: UserTrialSubscriptionRepository = {
+    featureFlag: "true",
     getUserSubscriptionDetail: () =>
       TE.right({
         state: SubscriptionStateEnum["ACTIVE"],
       }),
   };
-
-  // TODO
-  const trialSystemFeatureFlag = "true";
 
   const logger = {
     format: L.format.simple,
@@ -67,7 +65,6 @@ describe("SetWalletInstanceStatusHandler", () => {
       input: req,
       inputDecoder: H.HttpRequest,
       logger,
-      trialSystemFeatureFlag,
       userRepository,
       userTrialSubscriptionRepository,
       walletInstanceRepository,
@@ -95,7 +92,6 @@ describe("SetWalletInstanceStatusHandler", () => {
       input: req,
       inputDecoder: H.HttpRequest,
       logger,
-      trialSystemFeatureFlag,
       userRepository,
       userTrialSubscriptionRepository,
       walletInstanceRepository,
@@ -129,7 +125,6 @@ describe("SetWalletInstanceStatusHandler", () => {
       input: req,
       inputDecoder: H.HttpRequest,
       logger,
-      trialSystemFeatureFlag,
       userRepository,
       userTrialSubscriptionRepository,
       walletInstanceRepository,
@@ -163,7 +158,6 @@ describe("SetWalletInstanceStatusHandler", () => {
       input: req,
       inputDecoder: H.HttpRequest,
       logger,
-      trialSystemFeatureFlag,
       userRepository,
       userTrialSubscriptionRepository,
       walletInstanceRepository,
@@ -202,7 +196,6 @@ describe("SetWalletInstanceStatusHandler", () => {
       input: req,
       inputDecoder: H.HttpRequest,
       logger,
-      trialSystemFeatureFlag,
       userRepository,
       userTrialSubscriptionRepository,
       walletInstanceRepository,
@@ -236,7 +229,6 @@ describe("SetWalletInstanceStatusHandler", () => {
       input: req,
       inputDecoder: H.HttpRequest,
       logger,
-      trialSystemFeatureFlag,
       userRepository,
       userTrialSubscriptionRepository,
       walletInstanceRepository,
@@ -276,7 +268,6 @@ describe("SetWalletInstanceStatusHandler", () => {
       input: req,
       inputDecoder: H.HttpRequest,
       logger,
-      trialSystemFeatureFlag,
       userRepository: userRepositoryThatFailsOnGetUser,
       userTrialSubscriptionRepository,
       walletInstanceRepository,
@@ -319,7 +310,6 @@ describe("SetWalletInstanceStatusHandler", () => {
       input: req,
       inputDecoder: H.HttpRequest,
       logger,
-      trialSystemFeatureFlag,
       userRepository,
       userTrialSubscriptionRepository,
       walletInstanceRepository: walletInstanceRepositoryThatFailsOnBatchPatch,
@@ -356,7 +346,6 @@ describe("SetWalletInstanceStatusHandler", () => {
       input: req,
       inputDecoder: H.HttpRequest,
       logger,
-      trialSystemFeatureFlag,
       userRepository,
       userTrialSubscriptionRepository,
       walletInstanceRepository,
