+++ apps/io-wallet-user-func/src/infra/http/handlers/get-current-wallet-instance-status.ts
@@ -4,14 +4,14 @@ import * as H from "@pagopa/handler-kit";
 import * as RTE from "fp-ts/lib/ReaderTaskEither";
 import { pipe } from "fp-ts/lib/function";
 
-import { logErrorAndReturnResponse } from "../response";
-import { foo } from "./token-auth";
+import { logErrorAndReturnResponse } from "../error";
+import { requireWhitelistedUserFromToken } from "../whitelisted-user";
 
 export const GetCurrentWalletInstanceStatusHandler = H.of(
   (req: H.HttpRequest) =>
     pipe(
       req,
-      foo,
+      requireWhitelistedUserFromToken,
       RTE.chainW(({ id }) => getCurrentWalletInstance(id)),
       RTE.map(WalletInstanceToStatus.encode),
       RTE.map(H.successJson),
