+++ apps/io-wallet-user-func/src/infra/http/handlers/get-nonce.ts
@@ -3,7 +3,7 @@ import * as H from "@pagopa/handler-kit";
 import * as RTE from "fp-ts/lib/ReaderTaskEither";
 import { pipe } from "fp-ts/lib/function";
 
-import { logErrorAndReturnResponse } from "../response";
+import { logErrorAndReturnResponse } from "../error";
 
 export const GetNonceHandler = H.of(() =>
   pipe(
