+++ apps/io-wallet-user-func/src/infra/trial-system/health-check.ts
@@ -3,6 +3,7 @@ import * as RTE from "fp-ts/lib/ReaderTaskEither";
 import * as TE from "fp-ts/lib/TaskEither";
 
 export interface TrialSystemHealthCheck {
+  featureFlag: string;
   healthCheck: () => TE.TaskEither<Error, boolean>;
 }
 
@@ -12,10 +13,14 @@ export const getTrialSystemHealth: RTE.ReaderTaskEither<
   },
   Error,
   true
-> = ({ trialSystemClient }) =>
-  pipe(
-    trialSystemClient.healthCheck(),
-    TE.flatMap((isHealth) =>
-      !isHealth ? TE.left(new Error("trial-system-error")) : TE.right(isHealth),
-    ),
-  );
+> = ({ trialSystemClient: { featureFlag, healthCheck } }) =>
+  featureFlag === "true"
+    ? pipe(
+        healthCheck(),
+        TE.flatMap((isHealth) =>
+          !isHealth
+            ? TE.left(new Error("trial-system-error"))
+            : TE.right(isHealth),
+        ),
+      )
+    : TE.right(true);
