+++ apps/io-wallet-user-func/src/user.ts
@@ -5,8 +5,6 @@ import * as TE from "fp-ts/lib/TaskEither";
 import { pipe } from "fp-ts/lib/function";
 import * as t from "io-ts";
 
-import { Config } from "./app/config";
-
 export const User = t.type({
   id: NonEmptyString,
 });
@@ -50,6 +48,7 @@ const SubscriptionState = t.type({
 export type SubscriptionState = t.TypeOf<typeof SubscriptionState>;
 
 export interface UserTrialSubscriptionRepository {
+  featureFlag: string;
   getUserSubscriptionDetail: (
     userId: NonEmptyString,
   ) => TE.TaskEither<Error, SubscriptionState>;
@@ -61,30 +60,25 @@ export interface UserTrialSubscriptionEnvironment {
 
 const isUserSubscriptionActive: (
   userId: NonEmptyString,
-) => RTE.ReaderTaskEither<
-  {
-    trialSystemFeatureFlag: Config["trialSystem"]["featureFlag"];
-  } & UserTrialSubscriptionEnvironment,
-  Error,
-  boolean
-> =
+) => RTE.ReaderTaskEither<UserTrialSubscriptionEnvironment, Error, boolean> =
   (userId) =>
-  ({ trialSystemFeatureFlag, userTrialSubscriptionRepository }) =>
-    trialSystemFeatureFlag === "true"
+  ({
+    userTrialSubscriptionRepository: { featureFlag, getUserSubscriptionDetail },
+  }) =>
+    featureFlag === "true"
       ? pipe(
           userId,
-          userTrialSubscriptionRepository.getUserSubscriptionDetail,
+          getUserSubscriptionDetail,
           TE.map(({ state }) => state === "ACTIVE"),
         )
       : TE.right(true);
 
-// nome
-export const checkUserSubscription: ({ id }: User) => RTE.ReaderTaskEither<
-  {
-    trialSystemFeatureFlag: Config["trialSystem"]["featureFlag"];
-  } & UserTrialSubscriptionEnvironment,
+export const ensureUserInWhitelist: ({
+  id,
+}: User) => RTE.ReaderTaskEither<
+  UserTrialSubscriptionEnvironment,
   Error,
-  undefined
+  void
 > = ({ id }) =>
   pipe(
     id,
